Refactored Folder Sturcture improved code quality
