+++ src/components/AboutUI/aboutcomp.css
@@ -48,7 +48,7 @@ body {
     /* Layout Declrations */
     padding: 6em 0;
     /* Appearance Declrations */
-    background-image: url('../videos/cover.jpg');
+    background-image: url('../../videos/cover.jpg');
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
