+++ src/components/AirportAirlineUI/airlinesdata.jsx
@@ -1,18 +1,18 @@
-import QF from '../images/Airline/QF.jpg'
-import VA from '../images/Airline/VA.jpg'
-import JQ from '../images/Airline/JQ.jpg'
-import REX from '../images/Airline/REX.jpg'
-import QL from '../images/Airline/QL.jpg'
-import VAR from '../images/Airline/VAR.JPG'
-import ALI from '../images/Airline/ALI.jpg'
-import COB from '../images/Airline/COB.jpg'
-import PEL from '../images/Airline/PEL.jpg'
-import HIN from '../images/Airline/HIN.jpg'
-import LNK from '../images/Airline/LNK.jpg'
-import SHP from '../images/Airline/SHP.jpg'
-import SKT from '../images/Airline/SKT.jpg'
-import SKP from '../images/Airline/SKP.jpg'
-import AN from '../images/Airline/AN.jpg'
+import QF from '../../images/Airline/QF.jpg'
+import VA from '../../images/Airline/VA.jpg'
+import JQ from '../../images/Airline/JQ.jpg'
+import REX from '../../images/Airline/REX.jpg'
+import QL from '../../images/Airline/QL.jpg'
+import VAR from '../../images/Airline/VAR.JPG'
+import ALI from '../../images/Airline/ALI.jpg'
+import COB from '../../images/Airline/COB.jpg'
+import PEL from '../../images/Airline/PEL.jpg'
+import HIN from '../../images/Airline/HIN.jpg'
+import LNK from '../../images/Airline/LNK.jpg'
+import SHP from '../../images/Airline/SHP.jpg'
+import SKT from '../../images/Airline/SKT.jpg'
+import SKP from '../../images/Airline/SKP.jpg'
+import AN from '../../images/Airline/AN.jpg'
 
 export const AirlineData = [
     {
