+++ src/components/AirportAirlineUI/airportdata.jsx
@@ -1,48 +1,48 @@
-import SYD from '../images/Airport/SYD.jpg'
-import MEL from '../images/Airport/MEL.jpg'
-import BNE from '../images/Airport/BNE.jpg'
-import ADL from '../images/Airport/ADL.jpg'
-import PER from '../images/Airport/PER.jpg'
-import DAR from '../images/Airport/DAR.jpg'
-import HOB from '../images/Airport/HOB.jpg'
-import NEW from '../images/Airport/NEW.jpg'
-import CAI from '../images/Airport/CAI.jpg'
-import LAU from '../images/Airport/LAU.jpg'
-import GDC from '../images/Airport/GDC.jpg'
-import SCA from '../images/Airport/SCA.jpg'
-import CAN from '../images/Airport/CAN.jpg'
-import HAM from '../images/Airport/HAM.JPG'
-import PMQ from '../images/Airport/PMQ.jpg'
-import TWN from '../images/Airport/TWN.jpeg'
-import MCK from '../images/Airport/MCK.jpg'
-import KAR from '../images/Airport/KAR.jpg'
-import ROK from '../images/Airport/ROK.jpg'
-import ASP from '../images/Airport/ASP.jpg'
-import PHD from '../images/Airport/PHD.jpg'
-import GLD from '../images/Airport/GLD.JPG'
-import BAL from '../images/Airport/BAL.jpg'
-import BRM from '../images/Airport/BRM.jpg'
-import CFS from '../images/Airport/CFS.jpg'
-import NWM from '../images/Airport/NWM.jpg'
-import AYR from '../images/Airport/AYR.jpg'
-import WHT from '../images/Airport/WHT.jpeg'
-import ALB from '../images/Airport/ALB.jpg'
-import EMR from '../images/Airport/EMR.jpg'
-import KAL from '../images/Airport/KAL.jpg'
-import MIL from '../images/Airport/MIL.jpg'
-import WAG from '../images/Airport/WAG.jpg'
-import MIA from '../images/Airport/MIA.jpg'
-import ROM from '../images/Airport/ROM.jpg'
-import DUB from '../images/Airport/DUB.jpg'
-import PLH from '../images/Airport/PLH.jpg'
-import HRV from '../images/Airport/HRV.jpg'
-import BUN from '../images/Airport/BUN.jpg'
-import TAM from '../images/Airport/TAM.jpg'
-import ARM from '../images/Airport/ARM.jpg'
-import KIM from '../images/Airport/KIM.jpg'
-import MOR from '../images/Airport/LER.jpg'
-import LER from '../images/Airport/HRV.jpg'
-import GAM from '../images/Airport/GAM.jpg'
+import SYD from '../../images/Airport/SYD.jpg'
+import MEL from '../../images/Airport/MEL.jpg'
+import BNE from '../../images/Airport/BNE.jpg'
+import ADL from '../../images/Airport/ADL.jpg'
+import PER from '../../images/Airport/PER.jpg'
+import DAR from '../../images/Airport/DAR.jpg'
+import HOB from '../../images/Airport/HOB.jpg'
+import NEW from '../../images/Airport/NEW.jpg'
+import CAI from '../../images/Airport/CAI.jpg'
+import LAU from '../../images/Airport/LAU.jpg'
+import GDC from '../../images/Airport/GDC.jpg'
+import SCA from '../../images/Airport/SCA.jpg'
+import CAN from '../../images/Airport/CAN.jpg'
+import HAM from '../../images/Airport/HAM.JPG'
+import PMQ from '../../images/Airport/PMQ.jpg'
+import TWN from '../../images/Airport/TWN.jpeg'
+import MCK from '../../images/Airport/MCK.jpg'
+import KAR from '../../images/Airport/KAR.jpg'
+import ROK from '../../images/Airport/ROK.jpg'
+import ASP from '../../images/Airport/ASP.jpg'
+import PHD from '../../images/Airport/PHD.jpg'
+import GLD from '../../images/Airport/GLD.JPG'
+import BAL from '../../images/Airport/BAL.jpg'
+import BRM from '../../images/Airport/BRM.jpg'
+import CFS from '../../images/Airport/CFS.jpg'
+import NWM from '../../images/Airport/NWM.jpg'
+import AYR from '../../images/Airport/AYR.jpg'
+import WHT from '../../images/Airport/WHT.jpeg'
+import ALB from '../../images/Airport/ALB.jpg'
+import EMR from '../../images/Airport/EMR.jpg'
+import KAL from '../../images/Airport/KAL.jpg'
+import MIL from '../../images/Airport/MIL.jpg'
+import WAG from '../../images/Airport/WAG.jpg'
+import MIA from '../../images/Airport/MIA.jpg'
+import ROM from '../../images/Airport/ROM.jpg'
+import DUB from '../../images/Airport/DUB.jpg'
+import PLH from '../../images/Airport/PLH.jpg'
+import HRV from '../../images/Airport/HRV.jpg'
+import BUN from '../../images/Airport/BUN.jpg'
+import TAM from '../../images/Airport/TAM.jpg'
+import ARM from '../../images/Airport/ARM.jpg'
+import KIM from '../../images/Airport/KIM.jpg'
+import MOR from '../../images/Airport/LER.jpg'
+import LER from '../../images/Airport/HRV.jpg'
+import GAM from '../../images/Airport/GAM.jpg'
 
 export const AirportData = [
     {
