+++ src/components/HeroUI/HeroSection.css
@@ -1,5 +1,5 @@
 .hero-container {
-    background: url('../videos/blur2.jpg') center center/cover no-repeat;
+    background: url('../../videos/blur2.jpg') center center/cover no-repeat;
     height: 92vh;
     width: 100%;
     box-shadow: inset 0 0 0 1000px rgba(0, 0, 0, 0.2);
