+++ src/components/NavUI/footercomp.jsx
