+++ src/pages/about.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AboutComp from '../components/AboutComp';
+import AboutComp from '../components/AboutUI/AboutComp';
 
 export default function about() {
     return (
