+++ src/pages/reviews.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './review.css'
-import GetReviews from '../components/GetReview';
+import GetReviews from '../components/ReviewUI/GetReview';
 
 export default function Reviews() {
